proxy/requirements.txt
# Arkheia Enterprise Proxy -- Python dependencies
# No pinned versions for flexibility; CI pins via pip-compile.
# Minimum versions are tested working.

fastapi>=0.115.0
uvicorn[standard]>=0.31.1
httpx>=0.27.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
aiofiles>=23.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27.1

# Load testing (Phase 2)
# locust>=2.20.0
